feat(mbedtls): update to version 4.1.0

Closes IDF-15544

See merge request espressif/esp-idf!47234

+++ components/mbedtls/port/esp_timing.c
@@ -27,7 +27,7 @@ struct _hr_time
     struct timeval start;
 };
 
-unsigned long mbedtls_timing_get_timer( struct mbedtls_timing_hr_time *val, int reset )
+unsigned long long mbedtls_timing_get_timer( struct mbedtls_timing_hr_time *val, int reset )
 {
     struct _hr_time *t = (struct _hr_time *) val;
